Create shared util to get connector metadata
